% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{dat2df}
\alias{dat2df}
\title{Read data from KP single-point measurement}
\usage{
dat2df(datafile)
}
\arguments{
\item{datafile}{string with full path to datafile}
}
\value{
Dataframe with the following columns:
   $ point         : integer
   $ CPD           : mV     ## denoted WF in source file
   $ CPD.mean      : mV     ## denoted WF in source file
   $ CPD.sd        : mV     ## denoted WF in source file
   $ GD            : au
   $ GD.sd         : au
   $ steps         : au
   $ RH            : relative humidity %
   $ temperature   : degrees Celsius
   $ time          : seconds
   $ substrateid   : chr
   $ sampleid      : chr
}
\description{
Extracts data from Kelvin probe single-point measurements in *.dat (ASCII) files.
}

