% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fermifunction.R
\name{Fermi}
\alias{Fermi}
\title{The Fermi function}
\usage{
Fermi(energy, temperature = 0, fermilevel = 5.1)
}
\arguments{
\item{energy}{numeric vector, eV (must be supplied)}

\item{temperature}{numeric, temperature/Kelvin, default 0 K}

\item{fermilevel}{numeric, energy/eV NB: make sure this lies \emph{within} the range of the supplied energy vector. Defaults to Ef of gold, 5.100 eV.}
}
\value{
Dataframe with the following columns (and no extra attributes):
   $ energy           : numeric, the supplied energy range
   $ fermilevel       : numeric, the supplied fermi level
   $ temperature      : numeric, the supplied temperature
   $ fermi.occupied   : probability an electron state is occupied
   $ fermi.unoccupied : probability an electron state is unoccupied
}
\description{
The Fermi function
}
\examples{
\dontrun{Fermi(seq(1, 10, 0.1), temperature = 400, fermilevel = 5.5)}
}
