% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workfunction.R
\name{WF.sample}
\alias{WF.sample}
\title{Calculate WF of sample}
\usage{
WF.sample(CPD, WF.tip, version = "2")
}
\arguments{
\item{CPD}{measured CPD (in mV)}

\item{WF.tip}{the known WF of the tip (in mV)}

\item{version}{use version="2"}
}
\value{
depends on version, "2" returns a dataframe, "1" a scalar
}
\description{
Calculate WF of sample
}
\details{
use version="2", it calculates more data for you (mean, sd, etc.)
}

