% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workfunction.R
\name{WF.tip}
\alias{WF.tip}
\title{Calculate WF of tip}
\usage{
WF.tip(CPD, WF.gold = 5100)
}
\arguments{
\item{CPD}{measured contact potential difference value (in mV)}

\item{WF.gold}{the standard WF of gold (default 5100 mV)}
}
\value{
scalar, work function of tip
}
\description{
Calculate WF of tip
}
